HV: handle integral issue report by MISRA-C
The main focus on: constant suffix U/UL; parameters cast like
uint32 to a uint16 variable; unify some APIs interface,
consist with the callers.

also modify some places to unify code style

+++ hypervisor/arch/x86/guest/vmcall.c
@@ -149,7 +149,7 @@ int vmcall_vmexit_handler(struct vcpu *vcpu)
 	}
 
 out:
-	cur_context->guest_cpu_regs.regs.rax = ret;
+	cur_context->guest_cpu_regs.regs.rax = (uint64_t)ret;
 
 	TRACE_2L(TRACE_VMEXIT_VMCALL, vm->attr.id, hypcall_id);
