+++ hypervisor/include/arch/x86/irq.h
@@ -99,7 +99,7 @@ void vcpu_inject_extint(struct vcpu *vcpu);
 void vcpu_inject_nmi(struct vcpu *vcpu);
 void vcpu_inject_gp(struct vcpu *vcpu, uint32_t err_code);
 void vcpu_inject_pf(struct vcpu *vcpu, uint64_t addr, uint32_t err_code);
-void vcpu_make_request(struct vcpu *vcpu, int eventid);
+void vcpu_make_request(struct vcpu *vcpu, uint16_t eventid);
 int vcpu_queue_exception(struct vcpu *vcpu, uint32_t vector, uint32_t err_code);
 
 int exception_vmexit_handler(struct vcpu *vcpu);
